[FIX] project_timesheet_forecast: get allocated_hours of project
Before this commit, when the user goes to
`Planning > Schedule > By Project` and he has no access to Project
app or no access to one project linked to his shifts then an access
error is occurred. The reason is because the allocated hours of each
project linked to the shifts displayed in the gantt view to be able
to display the progress bar per project and so if the user cannot
access to at least one project then the access error is raised when
we try to fetch the allocated hours of that project.

Steps to reproduce the issue:
----------------------------
1. install `project_timesheet_forecast` module (with demo data)
2. log in as Marc Demo
3. go to `Planning > Schedule > By Project`

Actual behavior:
---------------
An access error is occurred because Marc Demo cannot read
`Research & Development` project and so the allocated hours cannot
be fetched for the progress bar.

Expected behavior:
-----------------
The user should see the gantt view and also the progress bar even if
he cannot access to at least one project linked to the shifts displayed
in the gantt view.

Related to the fix done in 16.0: #42869

closes odoo/enterprise#42886

X-original-commit: 968ad0c39488dc98d5f583769288ce9d0457351e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>