[FIX] project_timesheet_forecast: get allocated_hours of project
Before this commit, when the user goes to
`Planning > Schedule > By Project` and he has no access to Project
app or no access to one project linked to his shifts then the
allocated hours displayed will be 0 in the total of the progress bar
for that project. The reason is because the allocated hours of each
project linked to the shifts displayed in the gantt view to be able
to display the progress bar per project and so if the user cannot
access to at least one project then the allocated hours for that
project will be considered as 0 hour.

Steps to reproduce the issue:
----------------------------
1. install `project_timesheet_forecast` module (with demo data)
2. log in as Marc Demo
3. go to `Planning > Schedule > By Project`

Actual behavior:
---------------
The allocated hours of `Research & Development` will be equal to 0
for Marc Demo since he cannot access to that project.

Expected behavior:
-----------------
The user should see the right progress bar even if he cannot access to a
project linked to his shift.

closes odoo/enterprise#42869

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>