[IMP] web_grid: use keys shortcut for float_toggle_grid_cell
Before this commit, when the encoding is in Days in Timesheets,
the grid cell used will be the float_toggle one. If the user clicks
on a cell, he cannot use the keyboard shortcut to change the value
or move into another grid cell.

This commit allows to use the keyboard shortcut for
`float_toggle_grid_cell` and here is the different shortcut when the
grid cell is in edit mode:
- `space`: by hitting on space bar, the button of the float_toggle_grid_cell
  should be clicked to edit the value.
- `enter`: by hitting on enter key, the grid cell in edit mode will move
  into the grid cell below.
- `TAB`: by hitting on the tab key, the grid cell in edit mode will move
  into the next grid cell.

task-2618994

closes odoo/enterprise#42680

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>