[FIX] hr_appraisal: remove deadcode of KanbanView/openChat
OpenChat feature in `hr` is managed by the view widget
"hr_employee_chat", which is added in kanban template.
The override of Kanban View to add openChat feature is not called
at all, so this commit removes this dead code, and the js_class,
as its only purpose was adding openChat.

closes odoo/enterprise#42897

Related: odoo/odoo#125753
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>