[IMP] (website_)knowledge: clean frontend view files
Purpose:
--------
Since the refactoring of knowledge for portal users [1] and the one of
the sidebar [2], some code in knowledge was not used anymore, or only
relevant when website was installed.

For example, the "frontend" layout that was previously used by portal
and public users is now only used by public users. The favorite section
of this "frontend" layout is not relevant anymore since public users
cannot add articles to their favorites.

This commit renames some files, templates,... to ease the distinction
between what's used for portal users and what's used for public users.

It also cleans some methods and removes "dead" parts.

[1]: https://github.com/odoo/enterprise/commit/d66829ca7fcbf25bd4683963908d0973734abf96
[2]: https://github.com/odoo/enterprise/commit/b55ee8f06d7990fc71ee2430029775d5e8277020

Task-3346593

Part-of: odoo/enterprise#43304