[FIX] l10n_in_reports_gstr: _read_group in gst return period
When a user Creates/Open GST Return Period, the error comes like tuple indices
must be integers or slices, not str.

Steps to produce:
- Install the 'l10n_in_reports_gstr' module.
- Accounting > Reporting > India > GST Return Periods.
- Create a new record.
- Error will be produced.

Seems like this issue comes because of the new API of '_read_group' instead
of the 'read_group' method. It returns the tuple and the code expects
the dictionary.

Sentry-4262634781

closes odoo/enterprise#43146

X-original-commit: 5dc9d5cb243a48f9c2e431fed46ab31ca6971a41
Signed-off-by: Josse Colpaert <jco@odoo.com>