[FIX] timesheet_grid: prevent recursion due to timezone difference
Steps to reproduce:
-------------------
- go to user "My profile";
- change timezone to one where you are yesterday;
- reload the page;
- click on global start button;
- set a project;

Remark:
It is not possible to reproduce
this problem at any time of day.

Issue:
------
When we call the `action_add_time_to_timer` method,
we don't yet have an existing timer.
So we'll create one using the `action_timer_start` method (1).
This will start a timer for the current timesheet.
If the user does not yet have a timer in progress
(and we want to display the timer),
if the timesheet date is before today's date,
we will create a new timesheet with the context date
via `action_start_new_timesheet_timer`.
Then we start a new timer via `action_timer_start`.
We return to step (1).

Cause:
------
The problem is that we are comparing a date that takes into
account the Odoo context and a date that takes into account the system:
`self.date < fields.Date.today()`
can be written:
`fields.Date.context_today(self) < fields.Date.today()`

If the context is such that you find yourself one day ahead,
this condition will always be `True`.
We are therefore entering a recursion.

Solution:
---------
Use context to compare dates.

opw-3309214

closes odoo/enterprise#43015

X-original-commit: dc290c6e5c4369b65492efe532d8fc0eda7b7d0c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>