[FIX] pos_sale_stock_renting: remove dependency
A new module (`pos_sale_stock_renting`) has been introduced by [1]
but contains an error: the module depends on `point_of_sale` and
`sale_stock_renting`, but it uses a field (`sale_order_line_id`)
that comes from `pos_sale`. As a result, when the code is executed,
it generates a traceback because of the missing field.

This is the reason why [2] has been deployed: to avoid the traceback.
To do so, [2] changes the dependency `point_of_sale` into `pos_sale`.
But this does not really fix the situation: suppose a DB with the
module `pos_sale_stock_renting` installed before [2]. Then, the
update occurs and [2] is deployed. From now on, each time the system
tries to load the modules, it will have an issue when loading
`pos_sale_stock_renting`: it sees that a dependency is missing,
therefore it can not load the module, and it prints an error because
of that:
https://github.com/odoo/odoo/blob/9b0f7730d77e7da9cb08b30929f46dd4698ccd8e/odoo/modules/loading.py#L517-L518

[1] 09c474745a6ad356d01ab1a2bd7e4dbbaac12707
[2] ead75837661da9dada9dbdb510a44438715ce77c

sentry-3928883995

closes odoo/enterprise#42984

Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>