[FIX] l10n_mx_edi_pos: properly check country from company
In many pos l10n addons, we assume that country is present in the company, and
this is a wrong assumption. Thanks to the runbot error that we are able to
detect these mistakes in the code. In this commit, we fix the issue by checking
the presence of country first before actually accession the country code.

Runbot Error: 20917

closes odoo/enterprise#43019

X-original-commit: 75bc3975e8d1b7f1c0e5083144cef0a99a4750cf
Related: odoo/odoo#126117
Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>