[FIX] helpdesk: use `ticket_ref` instead of `id` in portal search
Steps to reproduce:
-------------------
- go to Settings / Technical / Sequences & Identifiers / Sequences;
- change Next Number for helpdesk.ticket;
- create a new ticket in helpdesk;
- go to helpdesk portal;
- select "Search in Reference";
- search for the new ticket reference.

Issue:
------
The ticket reference cannot be found.

Cause:
------
Since the commit: 16c9990d32ea6f8dcaf7c36dbb6183029f598fd0
The id of a ticket and its reference may diverge
(since the latter is based on the sequence).
But the search for a portal user is always based on the id.

Solution:
---------
Don't use the `id` field in the search for a portal user,
but use the `ticket_ref` field.

opw-3379346

closes odoo/enterprise#43124

X-original-commit: 8c4aea7977031079a8e55a5af29cd2854e9586c2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>