[FIX] knowledge: properties field should be well aligned with labels
When the user will add the property, the property field values and Add a
Property button are not well aligned with the property labels. Because in this
commit https://github.com/odoo/enterprise/pull/38074/commits/22ce8a389ac0a1e03b32077a20caff88fb1ebe39, padding was applied only in property labels.

Technical:
Adding margin to the labels and padding to the values and Add a Property button.
Now labels, values and Add a property button all are well aligned.

Task-3381693

closes odoo/enterprise#44930

X-original-commit: d6f311e9a31ff3e4c7b8ae43e88f6c17a9d632e6
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>