[FIX] timesheet_grid: find max of a list that contains date objects
Steps to reproduce:
-------------------
- being in a multi-company context;
- set "Lock Timesheets" setting on one of these companies;
- create timesheets for any employee in companies;
- validate these timesheets together;

Issue:
------
A traceback appears.

Cause:
------
For timesheets in a company with "Lock Timesheets" enabled,
employees will have the `last_validated_timesheet_date` field updated.
However, before searching for timesheets with a timer in progress,
we check whether at least one company has "Lock Timesheets" enabled (it does).
Therefore, when we search for timesheets with a timer in progress,
the record set `analytic_lines` contains all timesheets,
even those belonging to a company with "Lock Timesheets" deactivated.

The result is that we try to evaluate:
`max(analytic_lines.employee_id.sudo().mapped('last_validated_timesheet_date'))`
which can be:
`max([datetime.date, ..., False])`

Unfortunately, it is not possible to compare a date object with a boolean.

Solution:
---------
Filter the analytic lines to keep those which are in the case of a company
with "Lock Timesheets" enabled.
If any, perform the necessary operations on these timesheets only.

opw-3357524

closes odoo/enterprise#43046

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>