[FIX] website_helpdesk_knowledge: properly discard form changes
In [1], one tour which ends on a form view was forgotten (because it does not
end on a Knowledge form view but in Helpdesk), but the issue is the same:
If a tour ends on a form view with ongoing changes without being discarded, it
may cause `urgentSave` issues like mentionned in [2].

This commit aims to fix the runbot issue [3], in which a test running after the
tour being fixed failed because the ressources it relied upon were flushed by
the ongoing thread of the tour (the `helpdesk_team` record). This could result
in the `Help` menu button not being displayed in the navbar menu as it was
needed by the following test tour, ending in a failure of the test.

Add a check for a tour which add a file to send by message. It should also
ensure that the file finished being uploaded before ending the tour.

Modify the tour that sends a message so that it only ensures that the macro
writes correctly in the message body on the form dialog. The message sending
feature should be tested in mail.

[1]: https://github.com/odoo/enterprise/pull/42555
[2]: https://github.com/odoo/odoo/pull/96517
[3]: https://runbot.odoo.com/web/#id=21676&view_type=form&model=runbot.build.error&menu_id=405&cids=1

task-3383231

closes odoo/enterprise#43056

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>