[FIX] l10n_fr_fec_import: avoid setting taxes when importing the FEC
When importing the FEC with default taxes on some accounts, taxes may be
set on the account_move_lines created. This should not occur. In
addition, this will raise an exception as the tax_ids are inferred, then
the moves are posted, but then during `_post_process`, the partner_id is
set, triggering `_compute_all_tax` and resetting the taxes. Hence, we
get a "You cannot modify the taxes related to a posted journal item"
exception.

When creating the account_move_lines, give a value for `tax_ids` to
avoid the automatic setting.

opw-3359806

closes odoo/enterprise#43085

Signed-off-by: Quentin De Paoli <qdp@odoo.com>