[FIX] account_asset: don't create moves before lock date
Currently the asset depreciation moves computation does not take the
lock dates into account. It might thus create journal entries (in draft)
before the lock date, causing an error when trying to post them or
when you want to change the computation on the Asset form.

This fix checks whether there are any moves generated with a date before
the lock date and sets the date for these moves to the end of the first
period after the lock date (or containing the lock date if it's not the
last day of the period).

opw-3305506

closes odoo/enterprise#43081

Signed-off-by: William André (wan) <wan@odoo.com>