[FIX] account_followup: remove need of action on negative followup
Follow up reports have the "In need of action" state for partner with
unreconciled invoices which have a past due date. This is normal.
However, if a payment or credit note is made by/for a partner,
the Total Due for that partner might be 0 or negative but the followup
report stays in the "In need of action" state.
Because of this, clients were receiving reminders to pay a
negative or null value.

opw-3348941

closes odoo/enterprise#43813

X-original-commit: db7c302f3d2a49cc3be1bfc1e99fd6a6814a219d
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>