[FIX] account_followup: remove need of action on negative followup
Follow up reports have the "In need of action" state for partner with
unreconciled invoices which have a past due date. This is normal.
However, if a payment or credit note is made by/for a partner,
the Total Due for that partner might be 0 or negative but the followup
report stays in the "In need of action" state.
Because of this, clients were receiving reminders to pay a
negative or null value.

opw-3348941

closes odoo/enterprise#43090

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>