[FIX] industry_fsm_sale: `sale_line_id` shows SOs of child contacts
Current behaviour:
Sale Order Items on a task shows only SO related to the current
customer, or all SO if there is no customer on the task.

Expected behaviour:
If we have a customer on the task, show also SOs from child contact
of that customer (mirror the SO count in `Contacts` app)

Steps to reproduce:
- Install Contacts, Sales, Field Services
- Create 4 contacts, a company, with child contact A and B, a
  contact C who is child of A
- For each contact create an SO and confirm a field service product
- Now on any task, add the company as customer.
- In the Sale Order Item we can't select any of the SOs from the
  child contacts of the company.

Reason for the problem:
Since 184597b13dc5cd7d310539f087f4e60f641320e1 where we wanted to
show all SOs when no `partner_id` was set on the task, we made use
of `=?` operator which is `True` when there is no `partner_id`, but
it matched only for `=` when it's not the case, not taking into
account `child_of` of said contact.

Fix:
Extend the domain with an `OR` to take child contacts if the
`partner_id` is not `False`.

Affected versions:
- 16.0
- saas-16.1
- saas-16.2
- saas-16.3
- master

opw-3374191

closes odoo/enterprise#43093

Related: odoo/odoo#126285
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>