[MOV] knowledge: move the articles_structure content asset
This template has not the same usage as the others of its file, it is a
structural part of the Articles Structure Behavior that is used when it has been
mounted by OWL, not when it is in its "Blueprint" state. Therefore it makes more
sense to move it along with the OWL template of the Behavior, since it will be
used inside of it.

`knowledge_editor.xml` contains the `blueprint` that the `wysiwyg` need to
render in order for the `Behavior` engine to be able to mount the proper
`Behavior` components (which uses the OWL template of the component).

task-3383561

Part-of: odoo/enterprise#43087