[IMP] knowledge: rename Behavior templates consistently
Rename wysiwyg templates for Behavior anchors consistently.
There was a lot of displarity in the naming scheme of the `wysiwyg` templates
used as `blueprint` for Behaviors. Now, all of them are named as
`[BehaviorClassName]Blueprint`.

Rename `knowledge.article_structure` to `knowledge.ArticleStructureContent`.

Stop using a "modular" behavior type for the FileBehavior, since it will always
be the `FileBehavior` and it's not an abstract.

task-3383561

Part-of: odoo/enterprise#43087