[REM] purchase_enterprise: merge to purchase app
before this commit, purchase_enterprise module is
not adding any functionality to the system, other
than adding few demo data records and passing a
context to an existing action.

initially this module is added to bring the
dashboard functionality to the purchase app,
but with the introduction of the new spread sheet
app, the added dashboard is removed from
the purchase_enterprise app in this
commit: 61c8314

after this commit, this demo data and context
is moved to the purchase app and remove the
purchase_enterprise app.

closes odoo/enterprise#45986

Related: odoo/odoo#132475
Related: odoo/upgrade#5064
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>