[REF] knowledge: ensure that restoreSelection succeeded
Properly invalidate `Behaviors` that are being mounted for an obsolete state
of the htmlField. Since mounting Behaviors is asynchronous, the main Odoo Owl
app could patch the `HtmlField` while some `Behaviors` are not fully mounted
yet. At that point, those Behaviors should be destroyed or the mounting
process should be stopped. It was done before in a clunky way, and this commit
presents a proper way to do it:
  - The `htmlField` should maintain a validator object that will be attached
    to the `behaviorData` object during the mounting process.
  - Each time the `htmlField` changes state in a way that would make those
    Behaviors obsolete, it changes the value of a property in that object, and
    creates a new one that is valid and that will be given to the next wave of
    `behaviorData` objects.
  - Destroy obsolete Behaviors as soon as they are. Previously, some mounted
    obsolete Behaviors could be left hanging in the Handler until the user
    left the Knowledge Form view.

Enforce and verify return values for insertion and restoreSelection methods, to
properly invalidate and destroy a Behavior that could not be inserted during its
mounting process, because the HtmlField state changed in a way that makes it
impossible or indesirable to insert the Behavior.

task-3383561

Part-of: odoo/enterprise#43087