[FIX] web_studio: test reportEditorSidebar undeterministic crash
This commit follows df093649c016944744723ba4a58b974eafc3f1e4 which already tried to
make that test more stable.

Only it was partial as when we collapse/show a node in the report editor sidebar,
the active element of the bootstrap collapse framework does the reciprocal action.

That is:
- have a node active
- activate another one
- the active node hides AND the activated one shows.

This commit accounts for this by waiting for both events, as they are asynchronous in the bootstrap framework.

runbot-error-20792

closes odoo/enterprise#43139

closes odoo/enterprise#43436

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>