[IMP] knowledge: re-activate the view arch validation
Now that the forbidden `t-` owl directives have been removed from the
arch of the form view of Knowledge, we will re-enable the default arch
view validator. That way, we will be able to detect usage of forbidden
directives, attributes, etc and prevent potential issues.

Follow up of task: 3094058

task-3385229

closes odoo/enterprise#43168

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>