[FIX] web_grid: fonts of cells goes under the row header
Current behaviour:
The content of the cells are visible under the row header for uneven
row.

Expected behaviour:
The content of the cells shouldn't be visible under the row header
(where the project + task name is).

Steps to reproduce:
- Install Timesheets
- Open Timesheets App, switch the time scale to Month
- Scroll horizontally, observe that for the uneven rows, cells
  content are visible in the header.

Reason for the problem:
The row header has a higher z-index, so it's on top of the content
of the cells, but it has no background color, so we can see what is
behind it.

Fix:
Add a background color to the row, the same color as the background
color that is already there.

Affected versions:
- saas-16.2
- saas-16.3
- master

opw-3373381

X-original-commit: e6d1b9555d766b34f51d139f45c3cb1549ab5c91
Part-of: odoo/enterprise#43359