[FIX] web_grid: scale doesn't change when grouping by
Current behaviour:
If you apply a `group by` in the grid view, the scale is reset to
the default for the view.

Expected behaviour:
The scale of the view should be the same after a `group by`. (if we
set it as "month", it should stay as such, not switch back to "week")

Steps to reproduce:
- Install Timesheets
- Go to the Timesheets App
- Switch the scale to "month"
- Group by `project`
- Observe the scale button changed value, but the state of the grid
  itself didn't change.

Reason for the problem:
Before updating the props, we reset the current scale to the one
specified on the `arch` which is "week" by default.

Fix:
Remove this reset as it is not necessary.

Affected versions:
- saas-16.2
- saas-16.3
- master

opw-3373381

closes odoo/enterprise#43359

X-original-commit: 03f933219c6dd29f925659f5fe592bacccbbd636
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>