[FIX] planning: auto plan for recurrent shifts
Current behaviour:
Auto plan plans open shifts one by one. Also the highlights is
toggled on and off on each auto-plan, leading to un-highlighted
shifts that were planned each second time.

Expected behaviour:
All shifts in the current scale of the view should be planned
and all shifts should be highlighted to show what was auto-planned.

Steps to reproduce:
- Install Planning
- Make sure a role has multiple resources possible.
- Plan an open shift for 1 day, but is repeated each day until the
  end of the week.
- Click on autoplan, only the first shift is planned.

Reason for the problem:
A little 'oopsie' in the `_read_group` when get the max end date, it
was specified as `min`, when the intention (deduced from the variable
name) was to get the `max`. As for the highlight, we toggle the
currently highlighted shifts regardless of current highlights.

Fix:
Correct the `_read_group` for the planning, and for the highlight,
first toggle off the currently highlighted shifts, before
re-highlighting the newly planned ones.

Affected versions:
- saas-16.3
- master

opw-3375281

closes odoo/enterprise#43580

X-original-commit: 0600a49714b16004a59ca4bdefba12a072320e94
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>