[IMP] account_reports: Unrealized report
Before this PR the Unrealized report was not time dependant, since it was using
the amount_residual whatever when you make the payment and filter the report at
a certain date, the line could displayed wrong numbers.

Example behavior before the PR:
- Make a move at a date (2023-01-20), the rate of the currency change the next
day, so the move appears in the report
- Make a payment the next month (2023-02-05)
- Filter the report at the end of the month (2023-01-31)
- The line displayed has wrong number since the amount residual has changed with
the payment

So this PR make the report time dependant by using partial reconcile.
We have split the query into multiple parts. First we made a temporary table
that get the amount_residual and amount_currency at a certain date, the aml
id and currency_id for all moves that have a partial.

Then we use this temporary table, we add a condition to remove moves that are
fully paid at the date. And we will do a union all to get all move that don't
have a partial payment at a certain date.

Task-id: 3377112
X-original-commit: 84fea47d690c73e2b8e720979d32d2c0d1512354
Part-of: odoo/enterprise#44801
Co-authored-by: Florian Gilbert <flg@odoo.com>
Co-authored-by: Maximilien La Barre <malb@odoo.com>