[FIX] account_report: load more functionality
The issue:
When attempting to load additional records by clicking on 'load more...', we encounter a traceback. This occurs because, during the initial report load, the option next_progress is sent as a JSON string instead of a dictionary (or JSON object). Consequently, when accessing this object in the _get_aml_line function at line 612, a traceback is raised due to its string format.
Upon addressing this traceback, a new issue arises where the lines are correctly received from the backend but do not appear in the UI. This issue stems from the visibility of these lines not being updated, resulting in their absence from the user interface.

The fix:
- Sending the next_progress as a dictionary
- updating the visibility of the loaded lines

opw-3379697

closes odoo/enterprise#43888

X-original-commit: c5ee1355d99968643aabbb002feed6837ddbdc1b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>