[FIX] project: fix tasks overlap computation
Since https://github.com/odoo/enterprise/pull/26737, the task overlap computation ignores past tasks.
However, to determine which tasks are in the past, it compares the end date of the task, a timestamp in UTC,
with the "NOW()" sql function, whose timezone depends on that of the server, which can lead to problems if
the server is not in UTC.

To fix this, this commit forces the "NOW()" function to use the UTC timezone.

closes odoo/enterprise#43324

X-original-commit: b48e7377f08d91dc2da15d58312f72b6069af41d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>