[IMP] l10n_pl_reports: add export of JPK_VAT
Each month (or quarter), taxpayers have to give all the data about VAT to the Polish fiscal authorities.

We're only generating the xml for now (we don't send it directly yet).
We add a button on the tax report that would generate the data, in a new module so current users can have it.
All the new fields are added in community, because it will be useful for the EDI that will be done later.

We're currently generating JPK_V7M (2) and JPK_V7K (2).

To generate it, we fetch all the fields in SQL, as we could have a lot of data to generate (1 month or 1 quarter of operations).

task-3103658

closes odoo/enterprise#38166

Related: odoo/odoo#115225
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>