[FIX] account_reports_cash_basis: Wrong Computation of Matched Percentage
Main
-

Wrong Computation of Matched Percentage when there are journal items for
AR/AP for both the debit and credit.

Explanation
-

When a Journal Entry bears at least one journal time in the debit and
one journal item the credit. The computation of the `matched_percentage`
gets computated in the wrong way. This is because the balance value gets
ADDED prior to be converted into an ABSOLUTE value. The other way around
provide the proper values. That is, First convert the balance into an
ABS value and then ADD the values.

Have the following Journal Entry. It could not make sense but
demonstrate the issue.

**FY0**

|account|debit|credit|
|-|-:|-:|
|AR|350||
|AR||150|
|INCOME||200|

<img width="1149" alt="Screenshot 2023-06-27 at 11 17 34" src="https://github.com/odoo/odoo/assets/7598010/b2e54b5e-219a-4e68-9ead-f037d7df3187">

Reconciliations
-

Let us apply two reconciliations in two different fiscal years.

**FY1**

|account|debit|credit|
|-|-:|-:|
|BANK|350||
|AR||350|

<img width="1156" alt="Screenshot 2023-06-27 at 11 19 09" src="https://github.com/odoo/odoo/assets/7598010/7bc578dd-17aa-4842-9ea8-10f5c39e574b">

**FY2**

|account|debit|credit|
|-|-:|-:|
|AR|150||
|OTHER INCOME||150|

<img width="1164" alt="Screenshot 2023-06-27 at 11 26 51" src="https://github.com/odoo/odoo/assets/7598010/6dbd9867-1eda-46c2-88d5-f7110f64d153">

Current Behavior
-

```

CASE WHEN (aml.balance = 0 OR sub_aml.total_per_account = 0)
    THEN 0
    ELSE part.amount / ABS(sub_aml.total_per_account)
END as matched_percentage

...

ABS(SUM(balance)) AS total_per_account

```

The computation of `total_per_account` will yield 200

For zero fiscal  year:

![Screenshot 2023-06-27 at 11 39 38](https://github.com/odoo/odoo/assets/7598010/8272e5fb-8c10-4f5e-8da7-0a99ce4620b8)

For first reconciliation in FY1 the `matched_percentage` will be
(350 / 200) = 1.75 - That is 175% the income to be reported. 200 * 1.75 = 350 income

![Screenshot 2023-06-27 at 11 39 47](https://github.com/odoo/odoo/assets/7598010/3beeca97-3d85-4a34-b9d4-13d8c19194ec)

For second reconciliation in FY2 the `matched_percentage` will be:
For the first JE (150 / 200) = 0.75 - That is 75% the income to be reported. 200 * 0.75 = 150 income
For the third JE (150 / 150) = 1.00 - That is 100% the other income to be reported. 150 * 1.00 = 150 other income

![Screenshot 2023-06-27 at 11 39 58](https://github.com/odoo/odoo/assets/7598010/d579f58b-2296-4084-b290-c1158587758f)

**Conclusion**

At the end of the whole process you have reported:
-  350 + 150 in income (wrong) exceeds the amount of income to be reported at the End of the Cycle
- 150 in other income. (right)

Expected Behavior
-

```

CASE WHEN (aml.balance = 0 OR sub_aml.total_per_account = 0)
    THEN 0
    ELSE part.amount / ABS(sub_aml.total_per_account)
END as matched_percentage

...

SUM(ABS(balance)) AS total_per_account -- here goes the change ABS(SUM(balance)) -> SUM(ABS(balance))

```

The computation of `total_per_account` will yield 500

For zero fiscal  year:

![Screenshot 2023-06-27 at 12 02 33](https://github.com/odoo/odoo/assets/7598010/26273824-a9f1-40d1-a819-b971c0304dac)

For first reconciliation in FY1 the `matched_percentage` will be
(350 / 500) = 0.70 - That is 70% the income to be reported. 200 * 0.70 = 140 income

![Screenshot 2023-06-27 at 12 06 11](https://github.com/odoo/odoo/assets/7598010/cc50692c-b2a7-464f-8047-0c0989889abe)

For second reconciliation in FY2 the `matched_percentage` will be:
For the first JE (150 / 500) = 0.30 - That is 30% the income to be reported. 200 * 0.30 = 60 income
For the third JE (150 / 150) = 1.00 - That is 100% the other income to be reported. 150 * 1.00 = 150 other income

![Screenshot 2023-06-27 at 12 06 23](https://github.com/odoo/odoo/assets/7598010/24a41180-6bbd-4d2f-bd10-a2047b62a74c)

**Conclusion**

At the end of the whole process you have reported:
-  140 + 60 in income (right) the full amount of income to reported at the end of the cycle
- 150 in other income. (right)

closes odoo/enterprise#43478

X-original-commit: e8c78ddbfeb4bd5287ae08f0369baa786b439a97
Signed-off-by: William André (wan) <wan@odoo.com>