[FIX] helpdesk: fix computation of Hours Open in report
The field "Hours Open" in the ticket report is computed by extracting the hour
from the duration between the creation time and the closing time.

If that duration spans multiple days, the days are not taken into account, which
effectively clamps the value between 0 and 23.

To fix this, this commit converts the duration to seconds, then to hours.

Task-3386252

X-original-commit: 5cba304d18f4542b1c97b8a79e90831c1444c240
Part-of: odoo/enterprise#43265