[FIX] helpdesk: fix usage of "NOW()" in SQL views
In multiple places in the reporting SQL views, the function "NOW" is called without specifying a timezone.
Some fields will therefore be computed incorrectly if the timezone of the server is not UTC.

To fix this, this commit changes the calls to "NOW" to explicitely use the UTC timezone.

Task-3386252

closes odoo/enterprise#43297

X-original-commit: 6930493475885462fd6ecada448dfe76d4574419
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>