[FIX] hr_appraisal_survey: fix permission for non-survey user
You can set any employee as a manager of an appraisal and in the
appraisal you can ask for feedbacks. But, as long you don't have
any rights on surveys (or appraisal/admin) you can't open the answers
cause of a security issue.

Appraisal officer also have the right to access the survey result
as they could already have acces to it by just changing the manager
and then accessing them.

closes odoo/enterprise#43284

Task-id: 3366427
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>