[IMP] web_grid, timesheet_grid: adapt grid_view design to milk
Prior to this commit the gray that were used in timesheet were too dark
this change makes the view similar to the listview colors.

The grid_view cells were also too big which is why now its height is
reduced to 40px with the input taking the full available height.

The transitions have been removed. The project and task links now have
a body color by default and action color on hover.

Finally the `btn_timer` is now outline-secondary instead of primary
and is displaying a success state on active and hover.

Before this commit if we scrolled horizontally in the grid the row
containing the date had an extra column hidden behind the
`o_grid_navigation_buttons` before the row containing the time.

Since we can't account for the exact width of this element and it has to
adapt to the other titles in the next rows without changing the width of
the first column, custom SCSS had to be done.

task-3340318
part of task-3326263

Part-of: odoo/enterprise#45737