[FIX] global_filters: prevent unwanted errors while saving
In global filters of type 'relation', changing set domains of model X to
model Y, followed by changing the related model from X to Y generated a
field validation error while saving.

This issue occured because index of type `number` was passed to `add` method
whereas index of type `string` was passed to `delete` method. As a result,
'_wrongFieldMatchingsSet' was never emptied since 'delete' method did not
work as intended due to strict equality.

This commit fixes the issue by ensuring that type of index passed in
params of the methods is always `number`.

closes odoo/enterprise#49231

Task: 3272667
X-original-commit: 01ffeee134bc56f484dedc0432c7ca291d6bf644
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Rémi Rahir (rar) <rar@odoo.com>