[FIX] web_studio: differentiate identical nodes when moving a field
Steps to reproduce
==================

- Open a crm.lead form
- Open studio
- Switch to the "Extra Information" page
- Drag the "Company name" field somewhere else

-> The original field stays at the same place

Cause of the issue
==================

The simplified view is the following:

```xml
<form>
  <sheet>
    <group>
      <group name="lead_partner" attrs="{'invisible': [('type', '=', 'opportunity')]}">
        <field name="partner_name"/>
      </group>
    </group>
    <notebook>
      <page string="Extra Information" attrs="{'invisible': [('type', '=', 'lead')]}">
        <field name="partner_name"/>
      </page>
    </notebook>
  </sheet>
</form>
```

The following operation was send to the server

```py
{
    "type": "move",
    "target": {...},
    "position": "after",
    "node": {"tag": "field", "attrs": {"name": "partner_name"}},
}
```

Only the following attributes are use to differentiate fields:
['name', 'widget', 'options', 'display', 'optional'].
This is not enough. In this case, we wanted to move the second occurence
of the partner_name field, but the first one was actually moved.

Solution
========

Pass the xpath of the source field for the move operation.

opw-3270165

closes odoo/enterprise#43308

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>