[FIX] knowledge: ensure quasi-synchronous HtmlField `updateValue`
In Knowledge, the record and its latest changes should always be saved before
leaving the form view.

Summary of the changes in this commit:

- Add the `askChanges` call in `beforeLeave` of the Knowledge controller (this
  change is not done for every controller, because the `HtmlField` currently
  reports itself as dirty too often and that feature should be reworked. In
  Knowledge, loosing content of the article body is a greater concern than doing
  too many `save` calls, so we will enforce saving until the `dirty` process is
  reworked). This will allow to clean `askChanges` calls in the Knowledge
  module that were progressively added as monkey patches to try and make things
  work. The only `askChanges` that has to be kept is the one used when clicking
  on `/article` link, because that Behavior can be used outside of Knowledge.
- Ensure that the functions called during the `beforeunload` callback for the
  htmlField are quasi-synchronous (see community PR too).
- Introduce tests that verify those use cases

Technical explanation:

Normally, the htmlField.commitChanges is done when the blur event triggers
on the editable element, but in Knowledge, it is possible to call doAction
through a Behavior by clicking on an element inside the editable. This means
that the blur event is not triggered, and the record will not have been made
aware that the htmlField has pending changes, therefore the beforeLeave
callbacks of the Form view will not save these changes.

The issue lies with the current model implementation and more precisely the
`isDirty` method, which should ask fields for changes `askChanges` during the
`beforeLeave` call.

In case of `beforeUnload`, the callback provided will be called with a specific
`urgent` parameter and in those conditions, we should ensure that it is
quasi-synchronous, as the browser allocates a specific (short) amount of time
for the callback execution, and it is not enough to wait for an async rpc
roundtrip. Therefore we should not wait for `Behaviors` to be fully mounted
during the execution of the callback.

A consequence of that change is that we now have to change the way we remove
`Behavior` `anchor` children when we let Owl mount node inside it: we can remove
them synchronously when `Owl` finished its `mount` process instead of doing it
before the mounting starts (because the "after" occurs in the same microtask as
the node insertion by OWL, but the "before" does not). This means registering
those nodes before the mounting, and passing them as props to the `Behavior`
which will remove them when it is mounted (in `AbstractBehavior`). Following
this change, `blueprint` `Behavior` prop is replaced by `blueprintNodes` and its
usage is adjusted.

COMMUNITY PR: https://github.com/odoo/odoo/pull/127903

task-3372343

closes odoo/enterprise#44307

X-original-commit: eb7ea2c7814ecd15e62565ed9274ca34e2a016a6
Related: odoo/odoo#128949
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>