[FIX] web_gantt: use pointer events
Since [1], the draggable hooks in the views make use of pointer events
instead of mouse events to support touch devices. However the gantt view
was not entirely adapted to these changes (the draggable hook builder
was changed but the other listeners in the gantt view were not).

This commit properly switches all event listeners in the gantt view to
listen to pointer events.

[1] b85fdb60807a469d1457be7851ccd8068d680177

closes odoo/enterprise#43305

Related: odoo/odoo#126772
Signed-off-by: Géry Debongnie <ged@odoo.com>