[FIX] account_reports: multicurrency revaluation: don't require people to unfold lines manually to do the adjustment entry + generate one line per account
In 16.0, it was not possible to generate the adjustment entry anymore without manually unfolding every line of the report, which was obviously wrong. This was due to the fact the wizard used the report lines generated for model 'account.move.line' instead of 'account.account'. This error was shadowed in the tests, because the options passed to the report there forced the unfold_all key to True, causing the lines to always be present.

Also, because of the same error, the revaluation entry contained way too many move lines: one per line to adjust instead of aggregating them per account. We improve the tests so that this bug can be prevented in the future.

In the meantime, we stop directly putting the options as the context of the action opening the wizard to do it in a more proper (and sandboxed) way.

OPW 3271293

closes odoo/enterprise#43494

X-original-commit: 41bd17c6cc697a2739c388168179e592914519da
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>