[IMP] documents: improve tags following milk design
Before this commit, the tags color were hardcoded in the
'documents.facet' model and added as a style to the html
element which prevents, for example, the tags from changing
style when switching to dark mode.
This commit no longer uses style to define a tag color but
the 'o_tag_color_{index}' class instead. It avoids having
custom hexadecimal colors and keeps them in a unique css file.

Commit 5abb2de4ae6b048546afb82ab80aa14e473eb03a, which is part
of Milk design, introduced the tag design to documents.

Task-3384796

closes odoo/enterprise#43792

X-original-commit: e20a7aedcb0495214bee7f083ad717d912565691
Related: odoo/odoo#127677
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>